Prepare for Kexi 2.0
Compile and link fixes for:
- kexidb
- sqlite2, sqlite3 driver
- mysql driver
- postgresql driver



svn path=/trunk/koffice/; revision=536174

+++ kexi/kexidb/field.h
@@ -523,7 +523,7 @@ class KEXI_DB_EXPORT Field
 //<TMP>
 		/*! \return the hints for enum fields. */
 		Q3ValueVector<QString> enumHints() const { return m_hints; }
-		QString enumHint(uint num) { return (num < m_hints.size()) ? m_hints.at(num) : QString::null; }
+		QString enumHint(uint num) { return (num < (uint)m_hints.size()) ? m_hints.at(num) : QString::null; }
 		/*! sets the hint for enum fields */
 		void setEnumHints(const Q3ValueVector<QString> &l) { m_hints = l; }
 //</TMP>
